% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_PILE.R
\name{IV_PILE}
\alias{IV_PILE}
\title{Estimate an IV-optimal individualized treatment rule}
\usage{
IV_PILE(dt, kernel = "linear", C = 1, sig = 1/(ncol(dt) - 5))
}
\arguments{
\item{dt}{A dataframe whose first column is a binary IV 'Z', followed
by q columns of observed covariates, a binary
treatment indicator 'A', a binary outcome 'Y',
lower endpoint of the partial identification interval 'L',
and upper endpoint of the partial identification interval 'U'.
The dataset has q+5 columns in total.}

\item{kernel}{The kernel used in the weighted SVM algorithm. The user
may choose between 'linear' (linear kernel) and
'radial' (Gaussian RBF kernel).}

\item{C}{Cost of violating the constraint. This is the parameter C in
the Lagrange formulation.}

\item{sig}{Sigma in the Gaussian RBF kernel. Default is set to
1/dimension of covariates, i.e., 1/q. This parameter
is not relevant for linear kernel.}
}
\value{
An object of the type \code{wsvm}, inheriting from \code{svm}.
}
\description{
\code{IV_PILE} estimates an IV-optimal individualized treatment
rule given a dataset with estimated partial identification intervals
for each instance.
}
\examples{
\dontrun{
# It is necessary to install the package locClass in order
# to run the following code.

attach(dt_Rouse)
# Construct an IV out of differential distance to two-year versus
# four-year college. Z = 1 if the subject lives not farther from
# a 4-year college compared to a 2-year college.
Z = (dist4yr <= dist2yr) + 0

# Treatment A = 1 if the subject attends a 4-year college and 0
# otherwise.
A = 1 - twoyr

# Outcome Y = 1 if the subject obtained a bachelor's degree
Y = (educ86 >= 16) + 0

# Prepare the dataset
dt = data.frame(Z, female, black, hispanic, bytest, dadsome,
     dadcoll, momsome, momcoll, fincome, fincmiss, A, Y)

# Estimate the Balke-Pearl bound by estimating each constituent
# conditional probability p(Y = y, A = a | Z, X) with a multinomial
# regression.
dt_with_BP_bound_multinom = estimate_BP_bound(dt, method = 'multinom')

# Estimate the IV-optimal individualized treatment rule using a
# linear kernel, under the putative IV and the Balke-Pearl bound.


iv_itr_BP_linear = IV_PILE(dt_with_BP_bound_multinom, kernel = 'linear')
}

}
